maybe a useful connector
the main idea is to design the angled connector to have the most
possible strength by printing it on the side, as the most possible
strength comes from across the horizontal.

a 3d printed part is strongest across the horizontal plane, as
horizontal connections are printed together as a single strand of
plastic, vertical connections have less strength as they are stacks of strands.